requirements.txt
fastapi
uvicorn
gunicorn
scikit-surprise
Faker
pandas